requirements.txt
colormath==3.0.0
decorator==4.4.1
networkx==2.4
numpy==1.18.1
pycairo==1.19.0
